requirements.txt
# Web Framework
streamlit>=1.28.1
altair>=5.1.2
plotly>=5.17.0

# Core data stack (Python 3.12 compatible)
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
Pillow>=10.1.0
openpyxl>=3.1.2

# Computer vision
opencv-python-headless==4.10.0.84

# Backend API
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.31

# PDF handling
pymupdf>=1.24.4

# Additional dependencies for OMR processing
scikit-learn>=1.3.0  # For ML-based bubble classification (optional)
matplotlib>=3.7.0    # For plotting and visualization
seaborn>=0.12.0      # For enhanced data visualization
